+++ CHANGELOG.md
@@ -3,9 +3,11 @@
 ## [UNRELEASED] (NAME) - YYYY-MM-DD
 ### Added
 - Added support for rfc2622 ^+, ^-, ^n and ^n-m prefix operators
+- More tests for checking withdrawn hijacks
+
 
 ### Changed
-- TBD (Changed existing functionality)
+- Updated/optimized db query for removing withdrawn peers (newer announcement)
 
 ### Fixed
 - TBD (bug-fix)